update canoe and cats-effect ecosystem

+++ src/main/scala/dev/faustin0/bus/bot/domain/UserRepository.scala
@@ -2,5 +2,5 @@ package dev.faustin0.bus.bot.domain
 
 trait UserRepository[F[_]] {
   def create(user: User): F[Unit]
-  def get(id: Int): F[Option[User]]
+  def get(id: Long): F[Option[User]]
 }
